\name{amei-internal}
%% Part of the sensible export list but (currently?) documented as
%% internal (or waiting for documentation to be written):
%% <LIST>
\alias{PlotEpi}
\alias{PlotCosts}
\alias{PlotParams}
\alias{TimeSeriesOfDensity}
\alias{SimulateEpidemic}
\alias{VarStopTimePolicy}
\alias{SimulateManagementQuantiles}
\alias{SimulateEpidemicQuantiles}
\alias{mcmc.bknumu}
\alias{propose}
\alias{epi.log.post.b}
\alias{epi.log.post.k}
\alias{epi.gibbs.nu}
\alias{epi.gibbs.mu}
\alias{epimanage}
\alias{epi.update.prime}
\alias{epi.update.soln}
\alias{init.soln}
\alias{theta}
\alias{TimeSeriesOfDensities}
\alias{vaccinate}
\title{Internal amei Functions}
\description{
  Internal amei functions
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :)).
}
\keyword{ internal }
