\name{plot.optvac}
\alias{plot.optvac}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Plotting Vaccination Cost Surfaces }
\description{
  This function provides a visualization of the cost
  surface return by the \code{optvac} function and can
  show the best and worst vaccination strategy
}
\usage{
\method{plot}{optvac}(x, main = NULL, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ an \code{"optvac"}-class object }
  \item{main}{ and optional title for the plot; if not
  specified then a sensible default is generated automatically }
  \item{\dots}{ additional arguments passed to \code{\link{plot}}}
}
\details{
  The function documented here supports visualization of
  \code{"optvac"}-class objects which are generated by the
  \code{\link{optvac}} function
}
\value{
  The only output of this function is beautiful plots
}
\references{
\emph{A statistical framework for the adaptive management of 
epidemiological interventions} (2008). 
Daniel Merl, Leah R. Johnson, Robert B. Gramacy, and Marc S. Mangel.
Duke Working Paper 08-29.
\url{http://ftp.stat.duke.edu/WorkingPapers/08-29.html}
}
\author{
Daniel Merl <dan@stat.duke.edu>, 
Leah R. Johnson <leah@statslab.cam.ac.uk>, 
Robert B. Gramacy <bobby@statslab.cam.ac.uk>,
and Mark S. Mangel <msmangl@ams.ucsc.edu>
}
%\note{ ~~further notes~~ 
%
% ~Make other sections like Warning with \section{Warning }{....} ~
%}
\seealso{ \code{\link{optvac}}, \code{\link{getpolicy}} }
\examples{
## for an example of the usage of this function,
## please see the documentation for optvac as
## referenced in the See Also section, above
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ hplot }
