\name{amei-package}
\alias{amei-package}
\alias{amei}
\docType{package}
\title{Adaptive Management of Epidemiological Interventions}
\description{
  This package provides a flexible statistical framework for generating
  optimal epidemiological interventions.  The aim is to minimize the
  total expected cost of an emerging epidemic, while simultaneously
  propagating uncertainty regarding underlying disease parameters
  through to the decision process via Bayesian posterior inference.
  The strategies produced through this framework are adaptive:
  vaccination schedules are iteratively adjusted to reflect the
  anticipated trajectory of the epidemic given the current population
  state and updated parameter estimates.
}
\details{
\tabular{ll}{
Package: \tab amei\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-10-20\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Daniel Merl <dan@stat.duke.edu>, 
Leah R. Johnson <leah@statslab.cam.ac.uk>, 
Robert B. Gramacy <bobby@statslab.cam.ac.uk>,
and Mark S. Mangel <msmangl@ams.ucsc.edu>

Maintainer: 
Robert B. Gramacy <bobby@statslab.cam.ac.uk>
}
\references{
\emph{A statistical framework for the adaptive management of 
epidemiological interventions} (2008). 
Daniel Merl, Leah R. Johnson, Robert B. Gramacy, and Marc S. Mangel.
Duke Working Paper 08-29.
\url{http://ftp.stat.duke.edu/WorkingPapers/08-29.html}
}
\keyword{ package }
