\name{plot.optvac}
\alias{plot.optvac}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Plotting Vaccination Cost Surfaces }
\description{
  This function provides a visualization of the cost
  surface return by the \code{optvac} function and can
  show the best and worst vaccination strategy
}
\usage{
\method{plot}{optvac}(x, main = NULL, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ an \code{"optvac"}-class object }
  \item{main}{ and optional title for the plot; if not
  specified then a sensible default is generated automatically }
  \item{\dots}{ additional arguments passed to \code{\link{plot}}}
}
\details{
  The function documented here supports visualization of
  \code{"optvac"}-class objects which are generated by the
  \code{\link{optvac}} function
}
\value{
  The only output of this function is beautiful plots
}
\references{
D. Merl, L.R. Johnson, R.B. Gramacy, and M.S. Mangel (2010).
\dQuote{\pkg{amei}: An \R Package for the Adaptive Management of
Epidemiological Interventions}.  \emph{Journal of Statistical Software}
\bold{36}(6), 1-32. \url{http://www.jstatsoft.org/v36/i06/}
  
D. Merl, L.R. Johnson, R.B. Gramacy, M.S. Mangel (2009). \dQuote{A
Statistical Framework for the Adaptive Management of Epidemiological
Interventions}.  \emph{PLoS ONE}, \bold{4}(6), e5807.
\url{http://www.plosone.org/article/info:doi/10.1371/journal.pone.0005807}
}
\author{
Daniel Merl <danmerl@gmail.com>\cr
Leah R. Johnson <lrjohnson@uchicago.edu>\cr
Robert B. Gramacy <rbgramacy@chicagobooth.edu>\cr
and Mark S. Mangel <msmangl@ams.ucsc.edu>
}
%\note{ ~~further notes~~ 
%
% ~Make other sections like Warning with \section{Warning }{....} ~
%}
\seealso{ \code{\link{optvac}}, \code{\link{getpolicy}} }
\examples{
## for an example of the usage of this function,
## please see the documentation for optvac as
## referenced in the See Also section, above
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ hplot }
